pyproject.toml
[project]
name = "yamlin"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [{ name = "Bart Frenk", email = "bart.frenk@gmail.com" }]
requires-python = ">=3.10"
dependencies = ["pyyaml>=6.0.2"]

[project.scripts]
aero = "yamlin:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = ["devtools>=0.12.2"]
test = ["pytest>=8.3.4", "pytest-asyncio>=0.25.3"]

[tool.pytest]
asyncio_default_fixture_loop_scope = "function"
